changes per comments
