+++ src/routes/index.js
@@ -1,0 +1,2 @@
+export * from "./Register";
+export * from "./SignIn";
